refactor(graphql): plan candidateのfactoryを分ける
